backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
email-validator==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
resend==0.8.0